The space is not foster parented
```
A<table><tr> B</tr> </em>C</table>
```

The second space isn't foster parented. The `</tr>` triggers the
"anything else" clause of the in table text insertion mode which causes
` B` to be foster parented. The following space clears the pending table
character tokens list and thus the `</em>` inserts the space without
foster parenting.

This is also clear from the text node `A BC` in the result. A foster
parented second space would result in `A B C`.