Add testcase for open attribute value
This is a curious testcase because html5ever appears to be failing it.
At least piping it to html5ever like so appears to produce:

```
$ cargo run -p html5ever --example tokenize < ~/projects/html5gum/fuzz/mintest
    Finished dev [unoptimized + debuginfo] target(s) in 0.04s
     Running `target/debug/examples/tokenize`
ERROR: Bad character
TAG  : <d 0='&0=''>
CHAR : "\n"
OTHER: EOFToken

Tokenizer profile, in nanoseconds

       31248         total in token sink

       43594         total in tokenizer
       24126  55.3%  AttributeValue(Unquoted)
        7133  16.4%  Data
        4389  10.1%  TagOpen
        2986   6.8%  SelfClosingStartTag
        1363   3.1%  BeforeAttributeValue
        1283   2.9%  BeforeAttributeName
        1162   2.7%  TagName
        1152   2.6%  AttributeName
```

I could not find an older revision of the spec (in w3 or whatwg) that
would explain this behavior.

It's late and I'm tired, but I believe my reading of the spec is correct
and this should not emit a tag.