Add testcase: Incorrectly opened comment in script data
lol-html fails this testcase because its "tokenizer" aims to be too
smart and infers state change based on script data. It's a problem with
their test harness because they run the wrong layer of abstraction IMO.

I just sort of winged the line/col information, it would be good to get
those CI PRs merged so there's some sanity checks.