Move all CDATA section tests into one file
All other initial states are referenced outside the section 13.2.5
Tokenization[1] that describes the tokenizer state machine.

It's therefore reasonable to consider the 'CDATA section state'
an implementation detail and as such a tokenizer may choose to
avoid exposing it in it's public API.

In that scenario while all other tests could be integration tests
that solely use the public API, the tests for the 'CDATA section state'
would need to be performed internally rather than externally.

This commit moves all the 'CDATA section state' tests into a new
separate cdataSection.test file to faciliate such testing.

[1]: https://html.spec.whatwg.org/multipage/parsing.html#tokenization